requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
django-cors-headers==4.3.1
django-filter==24.3
Pillow==12.1.1